% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSamples_barplot.R
\name{compareSamples_barplot}
\alias{compareSamples_barplot}
\title{Compare the different conditions across samples}
\usage{
compareSamples_barplot(allData, controlName = "FS + FS", thPval = 0.05,
  res_stats = NULL)
}
\arguments{
\item{allData}{list with one element for each patient sample/cell line, there must be at least 2 runs (i.e. full cycle)
for each sample}

\item{controlName}{name of the control sample, default is "FS+FS"}

\item{thPval}{threshold on the p-value to be considered significant}

\item{res_stats}{output of function computeStatistics, to consider only significant conditions. Default is NA.}
}
\description{
\code{compareSamples_barplot}
}
\details{
This function compare the distribution of the values measures for each condition across 2 or more samples. 
For each cell line/patient it tests if each experimental condition is more promising than in the compared cell line/patient
and visualises with a barplot the conditions which are significantly different ordered from most to least promising (based on z-score)
}
\examples{
data("allData", package="BraDiPluS")
allData<-allData[c(1,2)]
res<-compareSamples_barplot(allData, controlName="FS + FS")
}
