% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStatistics.R
\name{computeStatistics}
\alias{computeStatistics}
\title{Compute and plot z-score and p-value.}
\usage{
computeStatistics(allData, controlName = "FS + FS", thPval = 0.1,
  thSD = NA, subsample = F, saveFiles = F)
}
\arguments{
\item{allData}{list with one element for each patient sample/cell line, there must be at least 2 runs (i.e. full cycle)
for each sample}

\item{controlName}{name of the control sample, default is "FS+FS"}

\item{thPval}{threshold on the p-value to be considered significant}

\item{thSD}{threshold on the maximum allows standard deviation of z-scores for each sample across runs. Default is NA, so not sample
is discarted based on this; can be set, e.g. to 1, to discart samples with high variability across runs.}

\item{subsample}{select TRUE to consider always (for all patient sample/cell line) the same number of runs when computing the combined p-value,
this makes points in the volcano plot more comparable. FALSE all runs available are considered.}

\item{saveFiles}{TRUE to save the plot as pdf. FALSE to plot it.}
}
\description{
\code{computeStatistics} compute and plot the z-score and the p-value for one or more patient sample/cell lines.
}
\details{
This function computes the z-score and the p-values for all patient samples/cell lines provided as argument. For each 
patient sample/cell lines, the z-scores are computed separately for each run and then averaged. P-values are also computed
separately for each run, using Wilcoxon rank sum test to verify if the response to the tested drugs (alone or in combination)
is significantly higher than the measurements for control condition (no perturbation). P-values were FDR-corrected for multiple
hypotheses testing and combined across different runs using Fisher’s method.
}
\examples{
data("allData", package="BraDiPluS")
allData<-list(patient3=allData[[5]])
res<-computeStatistics(allData, controlName="FS + FS", thPval=0.1, thSD=1, subsample=F, saveFiles=T)
}
