% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotData}
\alias{plotData}
\title{Plot the data.}
\usage{
plotData(data, channels, startTime = NA, endTime = NA, ymin = NA,
  ymax = NA, peaks = NA, samples = NA, ytext = (-0.005),
  savePlot = FALSE)
}
\arguments{
\item{data}{input data (formatted as data.frame with 4 columns: green, orange, blue, time)}

\item{channels}{a vector with the name of the channels we want to plot (e.g. channels=c("green", "blue")
to plot the green and the blue channel)}

\item{startTime}{initial time point to be plot. Default is NA, plots from the first time point in the data}

\item{ymin, ymax}{lower and upper limit of the y axes. Default is NA, for automatically defined range to include all data.}

\item{peaks}{peaks computed using the function \code{\link{peaksSelection}}}

\item{samples}{samples selected using the function \code{\link{samplesSelection}}}

\item{ytext}{position in the y axes of the sample names/numbers if samples argument is specified}

\item{savePlot}{if FALSE, the plot is not saved; if TRUE plot is saved with default file name; if string,
plot is saved with given filename. Plot is saved in .png format}
}
\description{
\code{plotData} allows to visualize the data and, possibly, the selected samples and peaks.
}
\details{
This function is used to visualize the desired range of data showing values for one or more channels. It can be
used also to plot the peaks selected on one channel using the function \code{\link{peaksSelection}} and/or the
samples identified using \code{\link{samplesSelection}}.
}
\examples{
data(BxPC3_data, package="BraDiPluS")
plotData(data=MyData, channels=c("blue", "green", "orange"))

res <- samplesSelection(data=MyData, BCchannel="blue",BCthr=0.01, BCminLength=100, distThr=16, plotMyData=F, barcodePos="before")
samples<-res$samples
plotData(data=MyData, channels=c("blue", "orange", "green"), samples = samples)

BCpeaks <- peaksSelection(MyData, channel="blue")
plotData(data=MyData, channels=c("blue", "orange", "green"), peaks = BCpeaks)
}
\seealso{
\code{\link{peaksSelection}} to select peaks in a specific channel, \code{\link{samplesSelection}} to separate
data in samples
}
