% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityAssessment.R
\name{qualityAssessment}
\alias{qualityAssessment}
\title{Remove outlier peaks.}
\usage{
qualityAssessment(runs)
}
\arguments{
\item{runs}{List with one element for each run. For each run the list element correspond to the peaks selected for the different
samples using \code{\link{selectSamplesPeaks}} or \code{\link{samplesSelection}}}
}
\value{
This function returns a list with the same structure of the one used as input but without the outlier peaks.
}
\description{
\code{qualityAssessment} removes low quality replicates.
}
\details{
This function allows to remove the peaks that are outliers based on the orange channel. Orange dye is added to the cells suspention
in order to monitor mixing of reagents in each plug. This function looks at the distribution of the 
values of all orange peaks and marks the possible outliers: corresponding peaks are removed from further
analysis. It can be done at one for different runs (where we call run a complete cycle of all the samples corresponding
to different experimental conditions).
}
\examples{
data(BxPC3_data,package="BraDiPluS")
res <- samplesSelection(data=MyData, BCchannel="blue",
BCthr=0.01, distThr=300, plotMyData=TRUE)
samples <-res$samples

# select the peaks for each sample
samplesPeaks <- selectSamplesPeaks(samples, channel="green", metric="median", baseThr=0.01, minLength=350, discartPeaks="first", discartPeaksPerc=5)

# remove outliers based on orange channel
runs<-list(run1=samplesPeaks)
runs.qa<-qualityAssessment(runs=runs) 
}
\seealso{
\code{\link{samplesSelection}} to select samples.
}
