% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSamples_barplot.R,
%   R/compareSamples_volcanoplot.R
\name{compareSamples_volcanoplot}
\alias{compareSamples_volcanoplot}
\alias{compareSamples_volcanoplot}
\title{Compare the different conditions across samples}
\usage{
compareSamples_volcanoplot(allData, controlName = "FS + FS", thPval = 0.1,
  thEffectSize = 0.3, parametric = T)

compareSamples_volcanoplot(allData, controlName = "FS + FS", thPval = 0.1,
  thEffectSize = 0.3, parametric = T)
}
\arguments{
\item{allData}{list with one element for each patient sample/cell line, there must be at least 2 runs (i.e. full cycle)
for each sample}

\item{controlName}{name of the control sample, default is "FS+FS"}

\item{thPval}{threshold on the p-value to be considered significant}

\item{thEffectSize}{threshold on the effect size}

\item{parametric}{TRUE to use parametric tests (t-test if 2 samples, ANOVA if more samples), FALSE to use non parametric
tests (Wilcoxon rank sum test if 2 samples, Kruskal-Wallis rank sum test if more than 2 samples).}

\item{allData}{list with one element for each patient sample/cell line, there must be at least 2 runs (i.e. full cycle)
for each sample}

\item{controlName}{name of the control sample, default is "FS+FS"}

\item{thPval}{threshold on the p-value to be considered significant}

\item{thEffectSize}{threshold on the effect size}

\item{parametric}{TRUE to use parametric tests (t-test if 2 samples, ANOVA if more samples), FALSE to use non parametric
tests (Wilcoxon rank sum test if 2 samples, Kruskal-Wallis rank sum test if more than 2 samples).}
}
\description{
\code{compareSamples_volcanoplot}

\code{compareSamples_volcanoplot}
}
\details{
This function compare the distribution of the values measures for each condition across 2 or more samples computing p-values
and effect size using parametric or non parametric tests. The comparison is visualised with a volcano plot showing the effect size vs the -log p-val

This function compare the distribution of the values measures for each condition across 2 or more samples computing p-values
and effect size using parametric or non parametric tests. The comparison is visualised with a volcano plot showing the effect size vs the -log p-val
}
\examples{
data("allData", package="BraDiPluS")
allData<-list(patient3=allData[[5]])
res<-compareSamples(allData, controlName="FS + FS")
data("allData", package="BraDiPluS")
allData<-list(patient3=allData[[5]])
res<-compareSamples(allData, controlName="FS + FS")
}
