% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutData.R
\name{cutData}
\alias{cutData}
\title{Cut a portion of the data.}
\usage{
cutData(data, startTime = NA, endTime = NA)
}
\arguments{
\item{data}{as imported from the .txt file saved by the in-house LabVIEW progam}

\item{startTime}{initial time point to be considered}

\item{endTime}{final time point to be considered}
}
\value{
A data frame with the same structure of the input data but limited to
the selected time range.
}
\description{
\code{cutData} returns a data structure limited to the selected time range.
}
\details{
This function allows to create a new data structure including only data in a user
defined time interval (e.g. to remove test measuremets or to select one
specific cycle).
}
\examples{
data(BxPC3_data, package="BraDiPluS")
MyDataNew <- cutData(data = MyData, startTime=6000, endTime=6500)
}
